Syntax error fixed

+++ app/Services/RemoveMenuItemService.php
@@ -3,6 +3,7 @@
 namespace App\Services;
 
 use App\Exceptions\ModelNotFoundException;
+use App\Exceptions\RestaurantNotApprovedException;
 use App\Models\Menu\Menu;
 use App\Models\Restaurant\Restaurant;
 
@@ -12,6 +13,9 @@ class RemoveMenuItemService
     {
         if ($menuItem->restaurant_id != $restaurant->id) {
             throw new ModelNotFoundException;
+        }
+        if ($restaurant->status !== 'approved') {
+            throw new RestaurantNotApprovedException;
         }
         $menuItem->delete();
     }
